Clean-up Benchmark_Context and hide implementation details
